Rename: led_icons.py => led_icon.py
