+++ finesse/gui/led_icon.py
